Make ParserInput Equatable.
Created unique IDs for each explicitly created ParserInput, instead of
using the source itself for comparing equality. Because of Generics
limitations we cannot store the collection itself in the struct, nor
can we store just the hashvalue because arrays are not hashable, or
store the hash value only if the source is hashable because "protocol
'Hashable' can only be used as a generic constraint because it has Self
or associated type requirements".