Better error reporting
Include the remainder in the error thrown. This
allows to lookup the position if the error is bubbled
all the way up. Also, <|> returns the error of the parser
that got the furthest. That produces possibly the best
error message.

Performance of the CSV suite is the same.

Moved some error types around, as not all are user errors,
some are programmer errors (e.g. negative count).